게시판 제목 불러오는 로직을 BoardPage -> BoardHeader로 이동
